ShaderUI : Move PlugAdder to GafferUI.PlugVisibilityGadget
It has the potential to be more generally useful. Here we're not changing any code or behaviour - just the name. Any behaviour changes will be made in separate commits.
