+++ Changes.md
@@ -6,6 +6,12 @@ API
 
 - DictPath : Added support for subclassing, without the need to override `copy()` or `_children()`.
 - PlugCreationGadget : Added counterpart to PlugCreationWidget, allowing plugs to be created easily in the GraphEditor.
+- PlugVisibilityGadget : Added gadget for managing plug visibility in the GraphEditor.
+
+Breaking Changes
+----------------
+
+- ShaderUI : Removed PlugAdder. Use `GafferUI.PlugVisibilityGadget` instead.
 
 1.7.0.0a5 (relative to 1.7.0.0a4)
 =========
